Move PreviouslyPublishedItem to models
